Rating The Wine Algorithm
This model is used to rate the wine quality on a scale of 1-10 - or, in other words, regressing the wine quality by training through a dataset using Sci-Kit Learn. This includes analyzing data through excel and pandas, manoeuvring the attributes, selecting the most appropriate model for the dataset and testing our results through cross validation.