An Automated Ethereum-based Trading Bot For Cryptocurrencies
An algorithm to predict the price of any cryptocurrency and hence trading for profits. This machine primarily uses the relative strength index on short time frames as an indicator to determine the position of the crypto market and makes decisions for the good of the owner.
This also includes collecting statistics of the current market data, applying TA-Lib indicators to completed candlesticks, and how to programmatically execute Ethereum orders based on these indicators using the Binance API key in real time.