Blockchain And DAPP Development For DeFi Tech
An algorithm to read and write the Ethereum blockchain network and scraping and analyzing the data for development of a decentralized application (DAPP) and essentially create a decentralized finance technology (DeFi). This program uses smart contracts to execute orders. I have preprogrammed it to read accounts through web3, receiving and sending transactions, reading, writing and deploying smart contracts and inspecting blocks.